+++ Plugins/BetterImageViewer/CHANGELOG.md
@@ -1,4 +1,8 @@
 # [BetterImageViewer](https://1lighty.github.io/BetterDiscordStuff/?plugin=BetterImageViewer "BetterImageViewer") Changelog
+### 1.0.2
+- Fixed plugin throwing a fit if image size was 1x1
+- Fixed some class issues making the authors name and send time not looking fancy
+
 ### 1.0.1
 - Nav buttons and info are now an overlay which means, fixed an issue from a certain trash plugin from a certain bad developer of who we do not speak of.